reorganized Admin components and started edit user functionality
