+++ bc-react/src/components/Admin/AdminUsers.js
@@ -1,28 +1,27 @@
 import React, {Component} from 'react';
-import AdminPlaceIndex from './AdminPlaceIndex';
-import AdminPlaceSearchBar from './Admin/AdminPlaceSearchBar';
-import PlaceStore from '../stores/PlaceStore';
-import AdminPlaceModal from './AdminPlaceModal';
-// import AdminPlaceModal from '../components/AdminPlaceModal';
+import AdminUserIndex from './AdminUserIndex';
+import AdminUserSearchBar from './AdminUserSearchBar';
+import AdminStore from '../../stores/AdminStore';
+import AdminUserModal from './AdminUserModal';
 //const api
 //only the most parent component should be responsible for fetching data
 
 //now in our Admin page we have users, and we want to put that into our Search Bar so it can use those props
 
-class AdminPlaces extends Component {
+class AdminUsers extends Component {
   constructor(props){
     super(props)
-    this.state = {places: PlaceStore.getPlaces(),
+    this.state = {users: AdminStore.getUsers(),
                   displayModal: false}
   }
-  updatePlaces(){
+  updateUsers(){
     this.setState({
-      places: PlaceStore.getPlaces() })}
+      users: AdminStore.getUsers() })}
 
   componentWillMount(){
-    PlaceStore.on('change', this.updatePlaces.bind(this)) }
+    AdminStore.on('change', this.updateUsers.bind(this)) }
 
-  showPlaceList(){
+  showUserList(){
     value: this.state.value }
 
   displayModal(){
@@ -30,25 +29,25 @@ class AdminPlaces extends Component {
 
   modalAdmin(){
     if(this.state.displayModal === true){
-    return (<AdminPlaceModal />)
+    return (<AdminUserModal />)
     } else { return ("") }}
 
   render(){
     return(
       <div id="admin_container">
-        <h3 className='center'>Places</h3>
+        <h3 className='center'>Users</h3>
         <div id="search_bar_wrapper">
           <button className="add_button" type="button"
             onClick={this.displayModal.bind(this)}>
-            + place </button>
-          {/* now SearchBar has access to places */}
-          <AdminPlaceSearchBar places={this.state.places}/>
+            + user </button>
+          {/* now SearchBar has access to users */}
+          <AdminUserSearchBar users={this.state.users}/>
         </div>
           <br></br><br></br>
-          <AdminPlaceIndex />
+          <AdminUserIndex />
           {this.modalAdmin()}
       </div>
       );
     }
   }
-export default AdminPlaces;
+export default AdminUsers;
